requirements.txt
aiofiles==0.5.0
aiohttp==3.7.4.post0
aiomysql==0.0.21
aiosqlite==0.17.0
aniso8601==7.0.0
asgiref==3.4.1
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==3.0.1
atlastk==0.12.10
attrs==21.2.0
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
bidict==0.21.3
black==21.9b0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.4.8
distlib==0.3.2
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.68.0
filelock==3.0.12
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
idna==3.2
iso8601==0.1.16
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
msgpack-python==0.5.6
multidict==5.1.0
mypy-extensions==0.4.3
orjson==3.6.3
pathspec==0.9.0
pipenv==2021.5.29
platformdirs==2.2.0
promise==2.3
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
PyMySQL==0.9.3
pypika-tortoise==0.1.1
python-dotenv==0.19.0
python-engineio==4.2.1
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.8.28
requests==2.26.0
Rx==1.6.1
scapy==2.4.5
six==1.16.0
socket.io-emitter==0.1.5.1
sqlparse==0.4.1
starlette==0.14.2
toml==0.10.2
tomli==1.2.1
tortoise-orm==0.17.7
typing-extensions==3.10.0.0
ujson==4.1.0
urllib3==1.26.6
uvicorn==0.13.4
virtualenv==20.7.2
virtualenv-clone==0.5.6
watchgod==0.7
websockets==8.1
yarl==1.6.3
